# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Agency Owner Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Deliverables ─────────────────────────────────────────────────────────
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: Not found
    patch:
      operationId: updatePayment
      tags: [payments]
      summary: Update a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentUpdate"
      responses:
        "200":
          description: Updated payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    delete:
      operationId: deletePayment
      tags: [payments]
      summary: Delete a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Documents ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, clientId, name, status, priority, progress, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        name:
          type: string
        status:
          type: string
          enum: [planning, waiting, design, development, testing, review, delivered, paused, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        startDate:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
        estimatedBudget:
          type: ["number", "null"]
        actualCost:
          type: ["number", "null"]
        revenue:
          type: ["number", "null"]
        profit:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        ownerNotes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            deliverables:
              type: array
              items:
                $ref: "#/components/schemas/Deliverable"
            documents:
              type: array
              items:
                $ref: "#/components/schemas/Document"
            notes:
              type: array
              items:
                $ref: "#/components/schemas/Note"
    ProjectInput:
      type: object
      required: [clientId, name, status, priority]
      properties:
        clientId:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [planning, waiting, design, development, testing, review, delivered, paused, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        startDate:
          type: string
        deadline:
          type: string
        estimatedBudget:
          type: number
        actualCost:
          type: number
        revenue:
          type: number
        description:
          type: string
        ownerNotes:
          type: string
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [planning, waiting, design, development, testing, review, delivered, paused, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        startDate:
          type: string
        deadline:
          type: string
        estimatedBudget:
          type: number
        actualCost:
          type: number
        revenue:
          type: number
        description:
          type: string
        ownerNotes:
          type: string

    # ─── Deliverable ─────────────────────────────────────────────────────────
    Deliverable:
      type: object
      required: [id, projectId, title, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        status:
          type: string
          enum: [pending, in_progress, review, done]
        deadline:
          type: ["string", "null"]
        assignedTo:
          type: ["string", "null"]
        completionDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Payment:
      type: object
      required: [id, clientId, invoiceNumber, amount, status, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        projectId:
          type: ["integer", "null"]
        invoiceNumber:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [paid, pending, overdue, cancelled]
        dueDate:
          type: ["string", "null"]
        paidDate:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
        remainingBalance:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [clientId, invoiceNumber, amount, status]
      properties:
        clientId:
          type: integer
        projectId:
          type: integer
        invoiceNumber:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [paid, pending, overdue, cancelled]
        dueDate:
          type: string
        paidDate:
          type: string
        paymentMethod:
          type: string
        remainingBalance:
          type: number
        notes:
          type: string
    PaymentUpdate:
      type: object
      properties:
        invoiceNumber:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [paid, pending, overdue, cancelled]
        dueDate:
          type: string
        paidDate:
          type: string
        paymentMethod:
          type: string
        remainingBalance:
          type: number
        notes:
          type: string

    # ─── Document ────────────────────────────────────────────────────────────
    Document:
      type: object
      required: [id, clientId, title, type, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        projectId:
          type: ["integer", "null"]
        title:
          type: string
        type:
          type: string
          enum: [contract, invoice, proposal, design, brand_assets, link, google_drive, github, figma, other]
        url:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Note:
      type: object
      required: [id, content, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: ["integer", "null"]
        projectId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        content:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...